Fix collapsed GUI render, harden install flow, purge dead code
GUI: GX2 uniform uploads must cover whole 16-byte constant slots. The
offset, scale, angle, blur and FXAA resolution counts had been set to
component counts, which the hardware drops silently, so the entire UI
drew inside one small box. Uploads now go through setUniform4 with the
value staged into four floats, slot alignment checked, and static
asserts pinning the assumptions.

Install flow: a WuxFlow state machine replaces three booleans; folder
cleanup, error and abort paths no longer leak or free the wrong object;
free space is checked before installing; the browser is torn down on
abort; the final note is shown once.

Safety: path traversal and TOC bounds checks, read results checked
instead of assumed, shader register copies clamped, allocation failures
logged rather than followed.

Dead code removed where each class was verified unreferenced:
ProgressWindow, AsyncExecutor, CursorDrawer, GuiImageAsync with the
game icon classes, and the unused 3D and fractal shaders.

+++ src/gui/FreeTypeGX.cpp
@@ -286,9 +286,9 @@ void FreeTypeGX::loadGlyphData(FT_Bitmap *bmp, ftgxCharData *charData)
 	uint16_t *dst = (uint16_t *)charData->texture->surface.image;
 	int32_t x, y;
 
-	for(y = 0; y < bmp->rows; y++)
+	for(y = 0; y < (int32_t)bmp->rows; y++)
 	{
-		for(x = 0; x < bmp->width; x++)
+		for(x = 0; x < (int32_t)bmp->width; x++)
 		{
 		    uint8_t intensity = src[y * bmp->width + x] >> 3;
             dst[y * charData->texture->surface.pitch + x] = intensity ? ((intensity << 11) | (intensity << 6) | (intensity << 1) | 1) : 0;
